+++ sample/src/modules/example-hello-world.ts
@@ -1,4 +1,3 @@
-import { ChannelType } from 'discord-api-types/v10';
 import { TextInputComponent } from 'discord.js';
 import { ChatCommand, ModalComponent, OptionBuilder, row } from 'purplet';
 
@@ -8,33 +7,11 @@ export default ChatCommand({
   descriptionLocalizations: {
     fr: 'Dit bonjour',
   },
-  defaultPermission: true,
-  options: new OptionBuilder()
-    .string('name', 'The name to say hello to', {
-      async autocomplete({ name }) {
-        return [
-          {
-            name: name ?? 'nice',
-            value: this.user.id,
-          },
-        ];
-      },
-    })
-    .string('word', 'How to say hello', {
-      choices: {
-        hello: 'Hello',
-        goodbye: 'Goodbye',
-        afternoon: 'Good Afternoon',
-      },
-    })
-    .channel('category', 'The category to say hello to', {
-      channelTypes: [ChannelType.GuildCategory],
-    })
-    .number('times', 'How many times to say hello', {
-      minValue: 1,
-      maxValue: 3,
-    })
-    .attachment('file', 'A category to say hello to'),
+  allowInDMs: true,
+  defaultPermission: false,
+  options: new OptionBuilder().string('name', 'The name to say hello to', {
+    minLength: 2,
+  }),
   async handle({ file }) {
     const modal = new Modal()
       .setTitle('Hello')
@@ -44,13 +21,6 @@ export default ChatCommand({
             .setCustomId('name')
             .setLabel('Name')
             .setPlaceholder('Name')
-            .setStyle('SHORT')
-        ),
-        row(
-          new TextInputComponent()
-            .setCustomId('cool')
-            .setLabel('Came')
-            .setPlaceholder('yeah')
             .setStyle('SHORT')
         )
       );